test(query-builder): reorganize order-by top-level into basic and sql-injection subfolders
