+++ test/functional/query-builder/order-by/basic/entity/Post.ts
@@ -1,7 +1,7 @@
-import { Entity } from "../../../../../src/decorator/entity/Entity"
-import { PrimaryGeneratedColumn } from "../../../../../src/decorator/columns/PrimaryGeneratedColumn"
-import { Column } from "../../../../../src/decorator/columns/Column"
-import { CreateDateColumn } from "../../../../../src/decorator/columns/CreateDateColumn"
+import { Entity } from "../../../../../../src/decorator/entity/Entity"
+import { PrimaryGeneratedColumn } from "../../../../../../src/decorator/columns/PrimaryGeneratedColumn"
+import { Column } from "../../../../../../src/decorator/columns/Column"
+import { CreateDateColumn } from "../../../../../../src/decorator/columns/CreateDateColumn"
 
 @Entity({
     orderBy: {
