--- test/functional/query-builder/order-by/entity/Comment.ts
@@ -1,22 +1,0 @@
-import { Entity } from "../../../../../src/decorator/entity/Entity"
-import { PrimaryGeneratedColumn } from "../../../../../src/decorator/columns/PrimaryGeneratedColumn"
-import { Column } from "../../../../../src/decorator/columns/Column"
-import { ManyToOne } from "../../../../../src/decorator/relations/ManyToOne"
-import { JoinColumn } from "../../../../../src/decorator/relations/JoinColumn"
-import { Post } from "./Post"
-
-@Entity()
-export class Comment {
-    @PrimaryGeneratedColumn()
-    id: number
-
-    @Column()
-    text: string
-
-    @Column({ name: "post_id" })
-    postId: number
-
-    @ManyToOne(() => Post)
-    @JoinColumn({ name: "post_id" })
-    post: Post
-}
