+++ test/functional/query-builder/order-by/sql-injection/entity/Post.ts
@@ -1,0 +1,29 @@
+import { Entity } from "../../../../../../src/decorator/entity/Entity"
+import { PrimaryGeneratedColumn } from "../../../../../../src/decorator/columns/PrimaryGeneratedColumn"
+import { Column } from "../../../../../../src/decorator/columns/Column"
+import { CreateDateColumn } from "../../../../../../src/decorator/columns/CreateDateColumn"
+
+@Entity({
+    orderBy: {
+        myOrder: "DESC",
+    },
+})
+export class Post {
+    @PrimaryGeneratedColumn()
+    id: number
+
+    @Column()
+    myOrder: number
+
+    @Column()
+    num1: number = 1
+
+    @Column()
+    num2: number = 1
+
+    @Column({ nullable: true })
+    title: string
+
+    @CreateDateColumn({ name: "created_at" })
+    createdAt: Date
+}
